make create_metadata_file public
