+++ src/loaders/compute_tools/tool_common.py
@@ -304,7 +304,7 @@ class ToolRunner:
                 for file in unzipped_files_to_delete:
                     os.remove(file)
 
-        _create_metadata_file(genomes_meta, batch_dir)
+        create_metadata_file(genomes_meta, batch_dir)
 
     def parallel_batch_execution(self, tool_callable: Callable[[Dict[str, GenomeTuple], Path, int, bool], None],
                                  unzip=False):
@@ -438,13 +438,18 @@ def run_command(command: List[str], log_dir: Path = None):
         raise ValueError(f'The command {command} failed with exit code {exit_code}')
 
 
-def _create_metadata_file(
-        meta: Dict[str, Dict[str, Union[str, Path]]],
-        batch_dir: Path
-):
-    # TODO: remove method once non-batched tool result parsers are implemented with the tool computation module
-    # create tab separated metadata file with tool generated genome identifier,
-    # original genome id and source genome file info.
+def create_metadata_file(
+        meta: Dict[str, Dict[str, str]],
+        batch_dir: Path,
+) -> None:
+    """
+    create tab separated metadata file with tool generated genome identifier,
+    original genome id and source genome file info.
+
+    :param meta: a dictionary of genome metadata information
+    :param batch_dir: the directory where the tool result files are stored
+    :return: None
+    """
 
     # create tool genome identifier metadata file
     genome_meta_file_path = os.path.join(batch_dir, loader_common_names.GENOME_METADATA_FILE)
