+++ test/src/loaders/compute_tools/tool_result_parser_test.py
@@ -1,0 +1,4 @@
+from src.loaders.compute_tools import tool_result_parser
+
+def test_noop():
+    pass
